known limitations & overview of non-optimized/optimized maps via mouse hovering
The optimized blocks are always displayed as the final overview.
To see the non-optimized blocks, the user can hover over the overview.
It then reveals the original map before optimization.

Blocks that have blue borders in the final overview are blocks from the
non-optimized map that were removed due to any kind issue from the
algorithm. -> this should be fixed in future commits